test: reorganize test files for better structure
- Consolidated 29 test files down to 15 for better organization
- Combined related test files:
  - test_evals.py: merged with test_evals_comprehensive.py
  - test_mcp_behavior.py: merged behavior and comprehensive tests
  - test_flakes.py: combined all flake-related tests
  - test_nixhub.py: consolidated nixhub integration tests
  - test_options.py: merged option info tests
  - test_channels.py: combined channel handling tests
  - test_integration.py: merged real and advanced integration
  - test_nixos_stats.py: combined stats regression tests
- Renamed test_server_comprehensive.py to test_server.py
- Renamed test_mcp_tools_comprehensive.py to test_mcp_tools.py
- Renamed test_fixes_comprehensive.py to test_regression.py
- Fixed missing imports (dataclass, patch) after file merging
- All test files now compile without syntax errors
